analysis.lami: correctly handle Number (double, long etc.) type graphing
LTTng analysis return mostly long value either for the time stamps or actual
measurements. SWTChart, on the other hand, deal only with "double" numerical
values. Casting long to double causes a loss of precision for big long value
(e.g time stamps). This loss of precision occurs when time
range with a big minimal value (2^62) but with little delta (1 ns) between
events exist. Graph generated for such time range would result in a single dot
which is a clear problem.

The presented solution uses linear mapping to preserve resolution
and precision when possible.

The linear mapping necessitates two range: the internal range and the external
range.

Each graph is responsible to provide an internal LamiGraphRange. This range is
the internal representation in double in which all raw value are mapped. For now the default internal
range is 0 to 1.

The external range [raw values range] is generated by finding the minimal value and maximal values
of aspects to be plotted.

Each point is then mapped to corresponding value from the internal range:

internalValue = (( externalValue - externalRange.minimum ) * ( internalRange.delta / external.delta )) + internalRange.minimum

Since the default internal range is from 0 to 1 al raw values are mapped to a
value from 0 to 1.

On graph tick generation axis formatter transform internal representation to
external representation and format the result.

Other change:
- Aspects now return their numerical values via resolveNumber to ensure no casting
is done.
- Move axis formatter to ui plugin.

Change-Id: I289180e10a7f1cbf6ecdd1beba93549b8fbe4c23
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>