tmf: Save a snapshot of the state when doing a trim operation
A "statedump" consisting of a full query at the start of the trim
range is done for every single state-system-driven analysis of this
trace.

This will ensure that the new trimmed trace does not lose any kind
of state information that could be inferred from the previous
events of the original trace.

Change-Id: Ia60585eaa117d75d45cc576ed39646fa1d0f3ad0
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>